Update main.4f7c3c3f8f54430c53d8.js
Users get a mixed-content error when they use the explorer with a node over http.  The community Pillar deeZNNutz.com is offering a trusted node to be used with this explorer.  This commit adds the following comment to the popup to add a node.

"A community run trusted node can be found at https://secure.deeznnodez.com:35997"